+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [PR #525](https://github.com/konpyutaika/nifikop/pull/525) - **[Operator/NiFiCluster]** Added support for IPv6 in NifiClutser
 - [PR #544](https://github.com/konpyutaika/nifikop/pull/544) - **[Helm Chart]** Added additionnal envs.
 - [PR #553](https://github.com/konpyutaika/nifikop/pull/553) - **[Helm Chart]** Added Service Account annotations and labels
+- [PR #555](https://github.com/konpyutaika/nifikop/pull/555) - **[Operator/NiFiDataflow]** Added displayName optional field
 
 ### Changed